Apply Brooks sweep and security feedback
